hkrc: added icons, fixed width for some columns
[issue tracker improvement]

Based on some HTML experimentation, fixed width columns could be
reliably achieved by using wrapper divs around cell contents. This
wrapping is done in a somewhat tricky way: since the
`Generator.get_postsummary_fields_flat` function returns a tuple of
functions which are called on the post items to generate the fields of
the output, we use a "maker function" to generate functions that take
a function, call it for the post item, then wrap it into the divs
required.

Important: right now, I'm stealing the icons from Canonical's
Launchpad. Even worse, I'm hotlinking them. This is not nice, and I
don't want to leave it like this, especially not in the next release.