hkweb: splitting posts added
[hkweb]

Holding Shift while clicking the "Add child" button opens the new post
next to its parent, and puts both into editing mode, thus allowing
for quickly splitting off parts of the text into a new post.

Note: right now, after saving, the parent and its new child are
still displayed in the table, next to each other. Reload the page to
reorganize the posts.

Note: very probably, when splitting a post, the new post should be a
sibling of the original instead of a child.