hkweb{,.js}: new root parsing errors are displayed
[new root]

If the text cannot be parsed as a post file, an error message is
presented to the user on the web interface, and xe can continue
editing and save again once the error is corrected.

If an error occurs, it is very probable that the user is trying to
post the body only. This could be avoided by opening the page with a
minimal template in place that shows how the raw post should look
like.