hkweb.js: removed unused parameter
[tidying]

The function `setPostContentRequest` had an unused parameter that is neither
used in the body of the function nor when calling it. It existed as a remnant
from an earlier mechanism that was used to discriminate between simultaneous
new-posts being edited on a single page (IIRC). This mechanism is not in use
any more, so I removed this parameter.