hkweb{,.js}: deleting posts implemented
[delete post]

Posts can now be deleted from hkweb. This change introduces a new
button and associated JS code plus an additional AJAX server to
actually perform the action on the server side.

TODO: if the root is deleted, the user has to face the disturbing
experience of seeing an empty thread page. Instead, the browser could
be redirected to the index page.