hkgen, hklib, hkweb: added rudimentary paged index
[paged index]

A new index generator is added to display the index in smaller chunks.
This is to reduce server and client loads.

As of now, this feature is very rudimentary, that is to mean
absolutely no UI support. The page does not display the number of the
current page, the number of available pages, no 'back' and 'forward'
links are available etc.

Another problem with this commit is that a great deal of code is being
duplicated, which is never a good thing.