hkweb{,.css}: added UI elements to paged index
[paged index]

This commit makes the paged index easily usable by adding the expected controls.

TODO: there is a visual problem: the table that contains the controls is always
centered, and the width of the cells can change. This means that the page
number is not precisely centered. This causes a visible problem on the first
and the last page.