hkweb{,.js}: save feature added to hkweb
[hkweb save]

This is the last step needed to actually deploy Hk on a server where the users
don't have shell access.

I am not really fond of the wording. 'Save changes to disk' does not
explicitely state that the saving will occur on the server side. Users may just
as well expect a diff file upon clicking a link labeled like this.